Refactor app to components
- Refactor application to components
- Create more readable directory structure
- Update file imports
