+++ src/components/Card/index.js
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const Card = styled.div`
+  box-shadow: 0 0 5px rgba(0, 0, 0, 0.1);
+  padding: 20px;
+  display: grid;
+  grid-template-columns: 150px 1fr;
+  grid-gap: 10px;
+  align-items: center;
+  font-weight: 100;
+`;
