+++ src/components/Container/index.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const Container = styled.div`
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+  grid-gap: 20px;
+  margin: 20px;
+  min-height: 60vh;
+`;
