+++ src/components/Footer/index.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const Footer = styled.div`
+  width: 98%;
+  border-top: 4px solid #388E3C;
+  padding: 10px;
+  margin: 20px auto;
+`;
