+++ src/components/Link/Link.js
@@ -1,0 +1,5 @@
+import styled from "styled-components";
+
+export const Link = styled.a`
+  text-decoration: none;
+`;
