+++ src/components/Wrapper/index.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const Wrapper = styled.div`
+  margin:0;
+  width:100%;
+  height: 100vh;
+`;
