+++ src/index.css
@@ -3,6 +3,7 @@ body {
   padding: 0;
   font-family: sans-serif;
 }
-*{
+
+* {
   box-sizing: border-box;
 }
